Fix new parent from within current hierarchy bug
Fixed bug where app would blow up when trying to give an item a new parent where the new parent was a child of the current element. Not sure how I didn't notice this until now. I still have to fix the issue, but now at least it should not corrupt the database.